Switched to bin64
* adjusted GUI functions to be flexible
* for computation there is no single point to change back

+++ inc/defines.hpp
@@ -22,6 +22,7 @@
 
 #include <complex>
 #include <string>
+#include <fftw3.h>
 
 #include "fixed_point.h"
 
@@ -44,7 +45,7 @@ typedef uint32_t meshindex_t;
 
 /* currently all of the below types have to be the same
  * use this to switch them all */
-typedef float data_t;
+typedef double data_t;
 
 typedef data_t csrpower_t;
 typedef std::complex<csrpower_t> impedance_t;
